+++ src/pages/QuestionPage/index.js
@@ -1,6 +1,5 @@
-import { useEffect, useState, useCallback } from 'react';
+import { useEffect, useState } from 'react';
 import { useParams } from 'react-router-dom';
-import useInfiniteScroll from '../../hooks/useInfiniteScroll';
 import getElapsedTime from '../../utils/getElapsedTime';
 import ReactionButtonBox from '../../components/ReactionButtonBox';
 import ModalQuestion from '../../components/ModalQuestion';
@@ -8,7 +7,7 @@ import QuestionPageContainer from './style';
 import QuestionHeader from '../../components/QuestionHeader';
 import getUserData from '../../api/getUserData';
 import QuestionCardHeader from '../../components/QuestionCardHeader';
-import getUserQuestionData from '../../api/getUserQuestionData';
+import useGetUserQuestions from '../../hooks/useGetUserQuestions';
 
 const QuestionItem = ({ user, question }) => {
   const [isFadedIn, setIsFadedIn] = useState(false);
@@ -66,38 +65,14 @@ const QuestionItem = ({ user, question }) => {
 const QuestionPage = () => {
   const { id } = useParams();
   const [user, setUser] = useState([]);
-  const [questions, setQuestions] = useState([]);
-  const [hasMore, setHasMore] = useState(true);
-  const [page, setPage] = useState(0);
+  const { questions, hasMore, elementRef, setQuestions } =
+    useGetUserQuestions(id);
   const [isShowModal, setIsShowModal] = useState(false);
   const [questionCount, setQuestionCount] = useState(0);
 
   const handleModalQuestion = () => {
     setIsShowModal(!isShowModal);
   };
-
-  const getUserQuestions = useCallback(async () => {
-    const delay = ms =>
-      new Promise(res => {
-        setTimeout(() => {
-          res();
-        }, ms);
-      });
-    await delay(500);
-
-    const responseQuestions = await getUserQuestionData(id, page);
-
-    if (responseQuestions.results.length === 0) {
-      setHasMore(false);
-    } else {
-      setQuestions(prevQuestions => [
-        ...prevQuestions,
-        ...responseQuestions.results,
-      ]);
-      setPage(prevPage => prevPage + 1);
-    }
-  }, [page, id]);
-  const elementRef = useInfiniteScroll(getUserQuestions);
 
   const fetchData = async () => {
     const responseUser = await getUserData(id);
